refactor(skills): remove duplicated code
